Add an example model tuning section to the template
In this example, just as a free-text global section. It could be
a subsection for each component, but it is likely that the tuning is
applied to the coupled system as a whole.

Additional tuning metadata could be added to the preamble is somehow
standardized across models. E.g. by defining broach tuning classes